Fix Wave 11: team-review follow-up (iterate revert + bug fixes)
Revert Base.iterate breaking change (NamedTuple → TSFrame) and fix
associated gaps, then address 5 bugs found in team review.

## Base.iterate restoration
- Restore iterate to return single-row TSFrame (user reverted manually)
- Add IteratorEltype=HasEltype + eltype(::Type{TSFrame})=TSFrame so that
  collect(ts) returns Vector{TSFrame} instead of Vector{Any}
- Add explicit Base.first(ts, n::Integer) override returning TSFrame
- Add explicit Base.last(ts, n::Integer) override (fixes MethodError from
  missing Base.keys definition)
- Update iterate tests to match TSFrame-based semantics

## Bug fixes
- getindex.jl: guard String indexing against non-Date/DateTime index
  (was MethodError, now ArgumentError with clear message) [MED-1]
- utils.jl: fix Base.summary writing to stdout instead of io arg [L7]
- to_period.jl: annotate n as Integer to reject Float inputs [L8]
- pctchange.jl: use copy() instead of collect() to preserve column
  storage type (PooledArray etc.) [M6]
- resample.jl: remove unreachable else branch in ffill/bfill and
  document the invariant; add :interpolate mixed-type behaviour to
  docstring [M5, M3]

+++ src/to_period.jl
@@ -40,7 +40,7 @@ for (fname, PType) in [
     (:to_microseconds, :Microsecond),
     (:to_nanoseconds,  :Nanosecond)
 ]
-    @eval function $fname(tsf::TSFrame, n=1)::TSFrame
+    @eval function $fname(tsf::TSFrame, n::Integer=1)::TSFrame
         n >= 1 || throw(ArgumentError("n must be >= 1, got $n"))
         ep = endpoints(tsf, $PType(n))
         TSFrame(tsf.coredata[ep, :], :Index; issorted = true, copycols = false)
